trash plant assets
Adds in the following assets:
Trash Plant
Trash Plant inside Dumpster
Closed Dumpster

These assets use a new color palette specifically for them. So they need a new toon shader material using the palette.

I will be dropping to my hands and knees and begging for Tati to create this material for me as my small, artist brain is incapable of processing the complexities of accomplishing this in Unity.